fix: the slidein effects need to be applied on all slides when used.
