+++ packages/core/src/components/QuickForm.tsx
@@ -1,5 +1,5 @@
 "use client";
-import React, { useMemo } from 'react';
+import React, { useMemo, useRef, useEffect, useState } from 'react';
 import { useQuickForm } from "../state/QuickFormContext";
 import { Ending, Submit, Intro, SlideRenderer } from "./index";
 import { mergeClasses } from '@griffel/react';
@@ -9,7 +9,26 @@ export type QuickFormProps = {
 }
 export const QuickForm: React.FC<QuickFormProps> = ({ className}) => {
     const { state, setIntroVisited } = useQuickForm();
-    const classes = useMemo(() => mergeClasses(className, state.classes.slide), [className, state.classes?.slide]);
+
+    const [slideEffectClassName, setSlideEffectClassName] = useState<string>();
+
+    const classes = useMemo(() => mergeClasses(className, state.classes.slide, slideEffectClassName), [className, state.classes?.slide, slideEffectClassName]);
+
+
+    let nextAllowedEffectTime = useRef(new Date().getTime());
+    useEffect(() => {
+        const timeout = setTimeout(() => {
+
+            setSlideEffectClassName(mergeClasses(state.classes.slideIsIn));
+
+        }, Math.max(150, nextAllowedEffectTime.current - new Date().getTime() + 150));
+
+        return () => {
+            clearTimeout(timeout);
+            setSlideEffectClassName(mergeClasses(state.classes.slideIsOut));
+            nextAllowedEffectTime.current = new Date().getTime() + 10;
+        }
+    }, [state.currIdx]);
 
     const renderComponent = () => {
         if (state.isIntroSlide && typeof state.data.intro !== "undefined") {
